Watch file events in the FS tree! :fire:
This ties together a bunch of prior work and is the last main chunk of work for the MVP of fomo.

I forked the notify crate so that a single channel could be used to communicate with the FS view to send it both user-defined queries and filesystem events. This avoids us having to set up an event forwarding thread and its associated overhead.